pyproject.toml
[project]
name = "iabot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "google-generativeai>=0.8.5",
    "llama-index>=0.12.43",
    "llama-index-embeddings-huggingface>=0.5.5",
    "llama-index-llms-gemini>=0.5.0",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot[webhooks]>=22.1",
    "scikit-learn>=1.7.0",
    "transformers>=4.52.4",
]
